Add RoomHelper for shared logic
