--- java/app/src/main/java/com/whereby/demoapp/Constants.java
@@ -1,11 +1,0 @@
-package com.whereby.demoapp;
-
-public class Constants {
-
-    /**
-     * Replace with your room URL.
-     * See https://docs.whereby.com/creating-and-deleting-rooms
-     */
-    public static String roomUrlString = "";
-
-}
