+++ java/app/src/main/java/com/whereby/demoapp/CustomWherebyRoomActivity.java
@@ -2,6 +2,6 @@ package com.whereby.demoapp;
 
 import com.whereby.sdk.WherebyRoomActivity;
 
-// Extend WherebyRoomActivity to customize the theme: set to `AppCompat.NoActionBar` in Manifest
+// Extend WherebyRoomActivity to customize the theme in Manifest
 // This allows to present the WherebyRoomActivity in full screen, for instance.
 public class CustomWherebyRoomActivity extends WherebyRoomActivity { }
