+++ java/app/src/main/java/com/whereby/demoapp/RoomActivityExample.java
@@ -1,26 +1,18 @@
 package com.whereby.demoapp;
 
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import android.content.BroadcastReceiver;
-import android.content.Context;
 import android.content.Intent;
-import android.content.IntentFilter;
 import android.os.Bundle;
-import android.util.Log;
 import android.widget.Button;
-
-import java.net.MalformedURLException;
-import java.net.URL;
 
 import com.whereby.sdk.*;
 
+import static com.whereby.demoapp.RoomHelper.*;
 import static com.whereby.sdk.WherebyConstants.*;
 
 public class RoomActivityExample extends AppCompatActivity {
-
-    private final static String TAG = RoomActivityExample.class.getSimpleName();
 
     private BroadcastReceiver mBroadcastReceiver;
 
@@ -37,9 +29,9 @@ public class RoomActivityExample extends AppCompatActivity {
     }
 
     @Override
-    protected void onStop() {
-        unregisterLocalBroadcastReceiver();
-        super.onStop();
+    protected void onResume() {
+        unregisterLocalBroadcastReceiver(mBroadcastReceiver, this);
+        super.onResume();
     }
     //endregion
 
@@ -61,48 +53,9 @@ public class RoomActivityExample extends AppCompatActivity {
         // Optional: this allows to receive async events through a broadcast receiver during the meeting.
         // Comment the following lines to disable.
         room.setEventBroadcastEnabled(true);
-        registerLocalBroadcastReceiver();
+        mBroadcastReceiver = registerLocalBroadcastReceiver(this);
 
         return room;
-    }
-
-    private URL createRoomUrl() {
-        URL roomURL = null;
-        try {
-            roomURL = new URL(Constants.roomUrlString);
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        }
-        return roomURL;
-    }
-
-    private WherebyRoomParameters createRoomParameters() {
-        WherebyRoomParameters roomParameters = new WherebyRoomParameters();
-        roomParameters.setDisplayName("Participant name");
-        //...
-        return roomParameters;
-    }
-
-    private void registerLocalBroadcastReceiver() {
-        IntentFilter intentFilter = new IntentFilter(ROOM_BROADCAST_EVENT_ACTION);
-        mBroadcastReceiver = new BroadcastReceiver() {
-            @Override
-            public void onReceive(Context context, Intent intent) {
-                WherebyRoomEvent event = (WherebyRoomEvent) intent.getSerializableExtra(ROOM_BROADCAST_EVENT_NAME);
-
-                // Process event:
-                Log.d(TAG, event.getRaw());
-            }
-        };
-        LocalBroadcastManager.getInstance(this).registerReceiver(mBroadcastReceiver, intentFilter);
-    }
-
-    private void unregisterLocalBroadcastReceiver() {
-        if (mBroadcastReceiver == null) {
-            return;
-        }
-        LocalBroadcastManager.getInstance(this).unregisterReceiver(mBroadcastReceiver);
-        mBroadcastReceiver = null;
     }
     //endregion
 }
